1 optimizer = 1 file + Abstraction super-module
