+++ docs/src/examples/DC-DC converter.jl
@@ -49,6 +49,7 @@ const ST = DI.System
 const SY = DI.Symbolic
 const CO = DI.Control
 const OP = DI.Optim
+const AB = OP.Abstraction
 
 # ### Definition of the system
 # we can import the module containing the DCDC problem like this 
@@ -65,7 +66,7 @@ hu = SVector(1)
 input_grid = DO.GridFree(u0, hu)
 
 using JuMP
-optimizer = MOI.instantiate(OP.Abstraction.Optimizer)
+optimizer = MOI.instantiate(AB.SCOTS.Optimizer)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("problem"), problem)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("state_grid"), state_grid)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("input_grid"), input_grid)
