+++ src/optim/abstraction/abstraction.jl
@@ -1,0 +1,7 @@
+module Abstraction
+
+include("SCOTS.jl")
+include("ellipsoids_abstractions.jl")
+include("lazy_ellipsoids_abstractions.jl")
+
+end
