+++ src/optim/optim.jl
@@ -1,6 +1,6 @@
 module Optim
 
-include("abstraction.jl")
+include("abstraction/abstraction.jl")
 include("bemporad_morari.jl")
 include("branch_and_bound.jl")
 include("q_learning.jl")
